DDF-04483 renaming user-notifications.view.js
