+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/user-notifications/user-notifications.js
